+++ OAuth/ResourceOwner/BitlyResourceOwner.php
@@ -38,6 +38,7 @@ class BitlyResourceOwner extends GenericOAuth2ResourceOwner
         parent::configureOptions($resolver);
 
         $resolver->setDefaults(array(
+            'use_bearer_authorization' => false,
             'authorization_url' => 'https://bitly.com/oauth/authorize',
             'access_token_url'  => 'https://api-ssl.bitly.com/oauth/access_token',
             'infos_url'         => 'https://api-ssl.bitly.com/v3/user/info?format=json',