+++ OAuth/ResourceOwner/EventbriteResourceOwner.php
@@ -12,7 +12,6 @@
 namespace HWI\Bundle\OAuthBundle\OAuth\ResourceOwner;
 
 use Buzz\Message\RequestInterface as HttpRequestInterface;
-use HWI\Bundle\OAuthBundle\Security\Core\Authentication\Token\OAuthToken;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 
 /**
@@ -28,6 +27,8 @@ class EventbriteResourceOwner extends GenericOAuth2ResourceOwner
     protected $paths = array(
         'identifier' => 'user.user_id',
         'nickname'   => 'user.first_name',
+        'firstname'  => 'user.first_name',
+        'lastname'   => 'user.last_name',
         'realname'   => array('user.first_name', 'user.last_name'),
         'email'      => 'email',
     );